refactor: updates UI to use common components
- separates form code into ImpersonateUserForm component
- makes use of existing SingleColumnView, CenteredContainer, StyledForm
- groups feature related code in a common folder
